move uri func to tbr
